Updated and Integrated Chart.js, removed references to ALL old charts.
The Food Waste Weight Performance Chart button (under personal bubble in Account) now links to the new Chart.JS file. The page functions as intended, showing a daily, weekly, and monthly view of food waste weight. The page could do with better styling and a note has been made to work on that. Furthermore, a dropdown of some description should be added to be able to switch to viewing food waste weight, to ghg, to cost. Again, a note has been made to add that.

All relevant routes to old charts have been removed pending a re-work of the database. In the current system, most uploaded data from all account types is going into a single collection "writtenFoodWasteData", this is untenable and a fix is on my TODO list, however, several other features and fixes have been prioritised and I have not had time to do both.