Fixed problems with UI rendering, reverted some changes.
Fixed some of the issues stemming from the tweaks made by Michael. The problem was created because the tweaks were aimed at the mobile UI and did not take into account the web view. This has been rectified.

One thing of note however is that the background image of blueberries is no longer present.