Fixed issue with FCM causing Android app to not work.
FCM is not compatible with a number of browser types, including the WebView used by the Android app. By adding a check to see if it is supported, and a null check if not, the webapp should now still load, although FCM will not work.

To use FCM with Android, we would need to migrate to React Native.