Home Page update
Changed the landing page to be concurrent with the mobile version. Removed the Navbar and Footer and with so, made sure that every link was still accessible via the account page and the styling of each page was not affected by the shift.
Created Copyright.js for when we need to have a copyright component.