Implemented Admin and Sub Accounts.
Admin accounts can now create Sub accounts, and said sub accounts submitted FoodWaste data goes to a collection associated with the Admin account.

This is still incomplete, as surplus and loss is not reflected as of yet.

Also, as of yet, there is no method for deleting a sub account or any user account. This is the next step.